+++ tests/test_contrib/test_hypothesis/test_laws/test_unsatisfiable_type.py
@@ -7,29 +7,24 @@ from returns.interfaces import mappable
 from returns.primitives.container import BaseContainer
 from returns.primitives.hkt import SupportsKind1
 
-pytestmark = pytest.mark.xfail(raises=AssertionError)
+pytestmark = pytest.mark.xfail(raises=RuntimeError)
 
 _ValueType = TypeVar('_ValueType')
 _NewValueType = TypeVar('_NewValueType')
 
 
-class _Wrapper(
-    BaseContainer,
-    SupportsKind1['_Wrapper', _ValueType],
+class _WithInitNoFlag(
+    SupportsKind1['_WithInitNoFlag', _ValueType],
     mappable.Mappable1[_ValueType],
 ):
-    _inner_value: _ValueType
-
-    def __init__(self, inner_value: _ValueType) -> None:
-        super().__init__(inner_value)
+    """Does not have any ways to be constructed."""
 
     def map(
         self,
         function: Callable[[_ValueType], _NewValueType],
-    ) -> '_Wrapper[_NewValueType]':
-        return _Wrapper(
-            'wrong-{0}'.format(function(self._inner_value)),  # type: ignore
-        )
+    ) -> '_WithInitNoFlag[_NewValueType]':
+        """We need `map` to have `laws`, should not be called."""
+        raise NotImplementedError
 
 
-check_all_laws(_Wrapper, use_init=True)
+check_all_laws(_WithInitNoFlag)