+++ returns/context/requires_context_ioresult.py
@@ -102,6 +102,8 @@ class RequiresContextIOResult(
 
     """
 
+    __slots__ = ()
+
     #: Inner value of `RequiresContext`
     #: is just a function that returns `IOResult`.
     #: This field has an extra 'RequiresContext' just because `mypy` needs it.