+++ CHANGELOG.md
@@ -12,6 +12,7 @@ See [0Ver](https://0ver.org/).
 
 - Enables Pattern Matching support for `Result` containers
 - Enables Pattern Matching support for `Maybe` container
+- Enables Pattern Matching support for `IOResult` container
 
 ## 0.16.0